Let `|indent` use `AsIndent`
