+++ askama/src/filters/mod.rs
@@ -22,8 +22,8 @@ mod urlencode;
 
 #[cfg(feature = "alloc")]
 pub use self::alloc::{
-    capitalize, fmt, format, indent, linebreaks, linebreaksbr, lower, lowercase, paragraphbreaks,
-    title, trim, upper, uppercase, wordcount,
+    AsIndent, capitalize, fmt, format, indent, linebreaks, linebreaksbr, lower, lowercase,
+    paragraphbreaks, title, trim, upper, uppercase, wordcount,
 };
 pub use self::builtin::{PluralizeCount, center, join, pluralize, truncate};
 pub use self::escape::{
@@ -32,7 +32,7 @@ pub use self::escape::{
 };
 pub use self::humansize::filesizeformat;
 #[cfg(feature = "serde_json")]
-pub use self::json::{AsIndent, json, json_pretty};
+pub use self::json::{json, json_pretty};
 #[cfg(feature = "urlencode")]
 pub use self::urlencode::{urlencode, urlencode_strict};
